yakshop/requirements.txt
django
djangorestframework
psycopg2
xmltodict
